final changes for project segmentación
